Remove deprecated snapshot generators and unused modules
