+++ README.md
@@ -91,11 +91,8 @@ These tools help measure closing line value (CLV) for your logged bets.
   in your `.env` to enable alerts. You can also define
   `DISCORD_ALERT_WEBHOOK_URL_2` to send the same CLV notifications to a second
   Discord channel.
-* **best_book_snapshot_generator.py** – captures the top EV opportunities across
-  sportsbooks. Set `DISCORD_BEST_BOOK_MAIN_WEBHOOK_URL` and
-  `DISCORD_BEST_BOOK_ALT_WEBHOOK_URL` in your `.env` to route main-line and
-  alternate-line snapshots to separate Discord channels.
-* **fv_drop_snapshot_generator.py** – highlights bets where the market probability ("Mkt %") increased compared to the previous snapshot. Set `DISCORD_FV_DROP_WEBHOOK_URL` to send the standard Bovada/BetOnline alerts. Define `DISCORD_FV_DROP_ALL_WEBHOOK_URL` to also post an all-books snapshot.
+* **dispatch_best_book_snapshot.py** – posts the best-priced opportunities across sportsbooks using the latest unified snapshot.
+* **dispatch_fv_drop_snapshot.py** – alerts when consensus probability rises since the previous snapshot.
 * **closing_odds_fetcher.py** – fetches closing odds for every game ID in your
   `logs/market_evals.csv` and writes a JSON file for the day. Useful after games
   finish if the monitor was not running.
