--- cli/replay_skipped_bets.py
@@ -1,5 +1,0 @@
-from cli.replay_pending_bets import main
-from core.bootstrap import *  # noqa
-
-if __name__ == "__main__":
-    main()
